Settings: notifications matrix becomes an inline section
The per-event channel matrix now renders as a /settings section like
Account/Slack (NotificationsSettingsBlock), fed by the shared settings
loader. /settings/notifications keeps the save action — the block's
fetcher POSTs there — and its GET redirects to /settings#notifications
so old links and the notification-email footer keep landing right.
